# WebAvancado/api/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"JsonObject": {
				"properties": {},
				"type": "object",
				"additionalProperties": {}
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "hello-express-mongo",
		"version": "1.0.0",
		"description": "basic API using Node and Express",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/api/user/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/user/getAll": {
			"get": {
				"operationId": "All",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/user/update": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"name": {
										"type": "string"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"name",
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/user/delete/{id}": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/kittyCat/create": {
			"post": {
				"operationId": "CreateCat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"species": {
										"type": "string"
									},
									"isFelv": {
										"type": "boolean"
									},
									"weight": {
										"type": "number",
										"format": "double"
									},
									"age": {
										"type": "number",
										"format": "double"
									},
									"color": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"species",
									"isFelv",
									"weight",
									"age",
									"color",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/kittyCat/getAll": {
			"get": {
				"operationId": "AllCat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/kittyCat/update": {
			"patch": {
				"operationId": "UpdateCat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"name": {
										"type": "string"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"name",
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/kittyCat/delete/{id}": {
			"delete": {
				"operationId": "DeleteCat",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/JsonObject"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}